Change book loading to the Books tab
